requirements.txt
alabaster==0.7.12
asn1crypto==1.0.1
attrs==19.3.0
Babel==2.7.0
backports.csv==1.0.7
biopython==1.74
bleach==3.1.0
bokeh==1.4.0
captcha==0.3
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-crispy-forms==1.7.2
django-dotenv==1.4.2
django-fernet-fields==0.6
django-import-export==1.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-tables2==2.1.1
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
ete3==3.1.1
idna==2.8
imagesize==1.1.0
importlib-metadata==1.3.0
ipython-genutils==0.2.0
jdcal==1.4.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter-core==4.6.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==5.0.3
nbsphinx==0.5.0
numpy==1.17.2
odfpy==1.4.0
openpyxl==3.0.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
Pillow==6.2.0
psycopg2==2.8.3
pudb==2019.2
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.5
PyQt5==5.13.1
PyQt5-sip==12.7.0
pyrsistent==0.15.6
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
simplejson==3.17.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.3.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
tablib==0.13.0
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
urwid==2.1.0
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
